requirements.txt
pyautogen==0.2.0
pydantic==1.10.8
yfinance>=0.2.31
matplotlib>=3.7.0
pandas>=2.0.0
numpy>=1.24.0